Add multi-channel volume support
AudioInputUSBOct::volume() now optionally takes an integer parameter between 1 and N where N is the number of channels. NOTE it is not 0-based! This reflects USB usage and the way most people think...

Without the optional parameter, the level set (by the host) for channel 1 is returned. Note also that this is now a static function, to match the static storage for the volume levels.

Added bool AudioInputUSBOct::isMuted(), does what it says on the tin.